XMLHttpRequest/setrequestheader-bogus-header.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: setRequestHeader() header argument checks</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function try_header(header) {
        assert_throws("INVALID_STATE_ERR", function() { client.setRequestHeader(header, "test") })
      }
      test(function() {
        var client = new XMLHttpRequest()
        client.open("GET", "...")
        try_header("")
        try_header("t<")
        try_header("t:t")
        try_header("t t")
        try_header(" t")
        try_header("\tt")
        try_header("t\rt")
        try_header("t\nt")
      })
    </script>
  </body>
</html>
